refactor: move html logic into packages
